+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [0.1.2] - October 6, 2020
+
+* Updated the [setAll], [setRange], and [replaceRange] methods to allow
+their supplied iterable to contain values already in the list if the resulting
+list doesn't contain any duplicate values once all values have been set.
+
 ## [0.1.1] - October 2, 2020
 
 * Set [nullable] to `true` by default in the [UniqueList.strict] factory constructor and the [toUniqueList] extension method.